Move compiler into eflang.ear.compiler package
