+++ ear-compiler/src/main/java/eflang/ear/compiler/EARPrograms.java
@@ -1,4 +1,4 @@
-package eflang.ear;
+package eflang.ear.compiler;
 
 public class EARPrograms {
     public static String test = "IN 0;IN 1;ADD @0 1;OUT 1;";
