+++ vibe/src/main/java/eflang/vibe/VibeController.java
@@ -2,7 +2,7 @@ package eflang.vibe;
 
 import eflang.core.Parser;
 import eflang.core.StringMusicSource;
-import eflang.ear.EARCompiler;
+import eflang.ear.compiler.EARCompiler;
 import eflang.ear.core.EARException;
 import eflang.ear.core.Scale;
 import eflang.ear.core.Scales;
